feat: useCredentialsの呼び出し側にOPS, CASの取得先情報を返す Signed-off-by: Yuuki Tsuchida <dev.yuukitsuchida@gmail.com>
