+++ apps/inspector/src/components/credentials/types.ts
@@ -9,12 +9,14 @@ import type {
   WebMediaProfile,
 } from "@originator-profile/model";
 import type {
+  CredentialSource,
   FetchCredentialSetResult,
   SourcedCredential,
 } from "@originator-profile/presentation";
 import type {
   IntegrityVerifyResult,
   VerifiedCas,
+  VerifiedOp,
   VerifiedOps,
 } from "@originator-profile/verify";
 
@@ -30,6 +32,30 @@ export type CredentialsProps = {
   wmp?: WebMediaProfile;
   framesCas: FramesVerifiedCas;
 };
+
+/**
+ * OP の取得経路。
+ * ページ埋め込み/外部URL(CredentialSource)に加え、
+ * ページ由来ではない経路(レジストリ・Site Profile)も表現する。
+ */
+export type OpOrigin =
+  | CredentialSource
+  | { kind: "registry" }
+  | { kind: "site-profile" };
+
+/** 出所(取得経路)付き検証済み Originator Profile */
+export type VerifiedOpWithSource = VerifiedOp & { source: OpOrigin };
+export type VerifiedOpsWithSource = VerifiedOpWithSource[];
+
+/**
+ * 出所(取得経路)付き検証済み Content Attestation。
+ * CAS はページ由来(ページ埋め込み/外部URL)以外の取得経路を持たないため、
+ * OPS の {@link OpOrigin} と異なり source は {@link CredentialSource} のみとなる。
+ */
+export type SupportedVerifiedCaWithSource = SupportedVerifiedCa & {
+  source: CredentialSource;
+};
+export type SupportedVerifiedCasWithSource = SupportedVerifiedCaWithSource[];
 export type FrameLocation = { origin: string; url: string };
 
 export type FetchCredentialsMessageResponse = FrameLocation & {
@@ -62,7 +88,7 @@ export type LinkVerificationResult = {
 };
 
 export type FrameVerifiedCas = FrameResponse &
-  FrameLocation & { cas: SupportedVerifiedCas };
+  FrameLocation & { cas: SupportedVerifiedCasWithSource };
 export type FramesVerifiedCas = FrameVerifiedCas[];
 
 export type ArticleLike =
